+++ .gitignore
@@ -107,9 +107,6 @@ fabric.properties
 
 ### VisualStudioCode ###
 .vscode/*
-!.vscode/tasks.json
-!.vscode/launch.json
-!*.code-workspace
 
 ### VisualStudioCode Patch ###
 # Ignore all local history of files
